+++ src/lib.rs
@@ -62,10 +62,10 @@ use serde::{Deserialize, Serialize};
 /// or parse an invalid repository spec
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub enum ParseError {
-    /// Returned by [`GHRepo::from_str`], [`GHRepo::from_url`], or
-    /// [`GHRepo::from_str_with_owner`] if given a string that is not a valid
-    /// GitHub repository URL or specifier; the field is the string in
-    /// question.
+    /// Returned by [`GHRepo::from_str`], [`GHRepo::from_url`],
+    /// [`GHRepo::from_str_with_owner`], or the `TryFrom<String> for GHRepo`
+    /// impl if given a string that is not a valid GitHub repository URL or
+    /// specifier; the field is the string in question.
     InvalidSpec(String),
 
     /// Returned by [`GHRepo::new`] or [`GHRepo::from_str_with_owner`] if given
@@ -95,6 +95,8 @@ impl error::Error for ParseError {}
 /// A `GHRepo` instance can be constructed in the following ways:
 ///
 /// - From an owner and name with [`GHRepo::new()`]
+/// - From a [`String`] of the form `{owner}/{name}` via the `TryFrom<String>`
+///   trait
 /// - From a GitHub URL with [`GHRepo::from_url()`]
 /// - From a GitHub URL or a string of the form `{owner}/{name}` with
 ///   [`GHRepo::from_str`]
@@ -109,7 +111,7 @@ impl error::Error for ParseError {}
 /// deserialized with the `serde` library.  Serialization produces a string of
 /// the form `{owner}/{name}`, and deserialization accepts any string of a form
 /// accepted by [`GHRepo::from_str`].
-#[derive(Clone, Debug, Eq, Hash, PartialEq)]
+#[derive(Clone, Eq, Hash, PartialEq)]
 pub struct GHRepo {
     fullname: String,
     slash_pos: usize,
@@ -261,8 +263,8 @@ impl GHRepo {
         format!("git@github.com:{}.git", self.fullname)
     }
 
-    /// Parse a GitHub repository URL.  The following URL formats are
-    /// recognized:
+    /// Parse a repository from a GitHub repository URL.  The following URL
+    /// formats are recognized:
     ///
     /// - `[http[s]://[<username>[:<password>]@]][www.]github.com/<owner>/<name>[.git][/]`
     /// - `[http[s]://]api.github.com/repos/<owner>/<name>`
@@ -282,9 +284,54 @@ impl GHRepo {
     }
 }
 
+impl From<GHRepo> for String {
+    /// Convert the repository to a string of the form `{owner}/{name}`
+    fn from(value: GHRepo) -> String {
+        value.fullname
+    }
+}
+
+impl fmt::Debug for GHRepo {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{:?}", self.fullname)
+    }
+}
+
 impl fmt::Display for GHRepo {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.fullname)
+        f.write_str(&self.fullname)
+    }
+}
+
+impl PartialEq<str> for GHRepo {
+    /// Compare the repository as though it were a string of the form
+    /// `{owner}/{name}`
+    fn eq(&self, other: &str) -> bool {
+        self.fullname == other
+    }
+}
+
+impl<'a> PartialEq<&'a str> for GHRepo {
+    /// Compare the repository as though it were a string of the form
+    /// `{owner}/{name}`
+    fn eq(&self, other: &&'a str) -> bool {
+        &self.fullname == other
+    }
+}
+
+impl std::ops::Deref for GHRepo {
+    type Target = str;
+
+    /// Dereference to a string of the form `{owner}/{name}`
+    fn deref(&self) -> &str {
+        &self.fullname
+    }
+}
+
+impl AsRef<str> for GHRepo {
+    /// Convert to a reference to a string of the form `{owner}/{name}`
+    fn as_ref(&self) -> &str {
+        &self.fullname
     }
 }
 
@@ -301,8 +348,36 @@ impl FromStr for GHRepo {
     /// specifier
     fn from_str(s: &str) -> Result<Self, ParseError> {
         match split_owner_name(s) {
-            Some((owner, name, "")) => GHRepo::new(owner, name),
+            Some((owner, _name, "")) => Ok(GHRepo {
+                fullname: s.to_owned(),
+                slash_pos: owner.len(),
+            }),
             _ => GHRepo::from_url(s),
+        }
+    }
+}
+
+impl TryFrom<String> for GHRepo {
+    type Error = ParseError;
+
+    /// Construct a `GHRepo` from a `String` of the form `{owner}/{name}`.
+    ///
+    /// Note that, unlike [`GHRepo::from_str()`], this trait does not accept
+    /// repository URLs.
+    ///
+    /// # Errors
+    ///
+    /// Returns a [`ParseError`] if `s` is not a valid repository specifier
+    fn try_from(s: String) -> Result<Self, ParseError> {
+        match split_owner_name(&s) {
+            Some((owner, _name, "")) => {
+                let slash_pos = owner.len();
+                Ok(GHRepo {
+                    fullname: s,
+                    slash_pos,
+                })
+            }
+            _ => Err(ParseError::InvalidSpec(s)),
         }
     }
 }
@@ -314,7 +389,7 @@ impl Serialize for GHRepo {
     where
         S: Serializer,
     {
-        serializer.collect_str(self)
+        serializer.serialize_str(&self.fullname)
     }
 }