+++ tests/ghrepo.rs
@@ -18,6 +18,24 @@ fn test_to_string() {
 }
 
 #[test]
+fn test_debug() {
+    let r = GHRepo::new("octocat", "repository").unwrap();
+    assert_eq!(format!("{r:?}"), r#""octocat/repository""#);
+    assert_eq!(format!("{r:#?}"), r#""octocat/repository""#);
+}
+
+#[test]
+fn test_misc_traits() {
+    let r = GHRepo::new("octocat", "repository").unwrap();
+    let s: &str = &r;
+    assert_eq!(s, "octocat/repository");
+    assert_eq!(r.as_ref(), "octocat/repository");
+    assert_eq!(r, "octocat/repository");
+    let s2 = String::from(r);
+    assert_eq!(s2, "octocat/repository");
+}
+
+#[test]
 fn test_api_url() {
     let r = GHRepo::new("octocat", "repository").unwrap();
     assert_eq!(
@@ -316,9 +334,80 @@ fn test_from_str(#[case] spec: &str, #[case] owner: &str, #[case] name: &str) {
     assert_eq!(GHRepo::from_str(spec), Ok(r));
 }
 
+#[rstest]
+#[case("jwodder/headerparser", "jwodder", "headerparser")]
+#[case("jwodder/none", "jwodder", "none")]
+#[case("nonely/headerparser", "nonely", "headerparser")]
+#[case("none-none/headerparser", "none-none", "headerparser")]
+#[case("nonenone/headerparser", "nonenone", "headerparser")]
+fn test_try_from_string(#[case] spec: String, #[case] owner: &str, #[case] name: &str) {
+    let r = GHRepo::try_from(spec).unwrap();
+    assert_eq!(r.owner(), owner);
+    assert_eq!(r.name(), name);
+}
+
 #[apply(bad_repos)]
 fn test_from_bad_str(#[case] spec: &str) {
     match GHRepo::from_str(spec) {
+        Err(ParseError::InvalidSpec(s)) if s == spec => (),
+        e => panic!("Got wrong result: {e:?}"),
+    }
+}
+
+#[apply(bad_repos)]
+#[case("git://github.com/jwodder/headerparser")]
+#[case("git://github.com/jwodder/headerparser.git")]
+#[case("git@github.com:jwodder/headerparser")]
+#[case("git@github.com:jwodder/headerparser.git")]
+#[case("GIT://GitHub.COM/jwodder/headerparser")]
+#[case("git@github.com:joe-q-coder/my.repo.git")]
+#[case("git@GITHUB.com:joe-q-coder/my.repo.git")]
+#[case("ssh://git@github.com/jwodder/headerparser")]
+#[case("ssh://git@github.com/jwodder/headerparser.git")]
+#[case("ssh://git@github.com/-/test")]
+#[case("SSH://git@GITHUB.COM/-/test")]
+#[case("https://api.github.com/repos/jwodder/headerparser")]
+#[case("http://api.github.com/repos/jwodder/headerparser")]
+#[case("api.github.com/repos/jwodder/headerparser")]
+#[case("https://api.github.com/repos/none-/-none")]
+#[case("HttpS://api.github.com/repos/none-/-none")]
+#[case("http://api.github.com/repos/none-/-none")]
+#[case("Http://api.github.com/repos/none-/-none")]
+#[case("Api.GitHub.Com/repos/jwodder/headerparser")]
+#[case("https://github.com/jwodder/headerparser")]
+#[case("https://github.com/jwodder/headerparser.git")]
+#[case("https://github.com/jwodder/headerparser.git/")]
+#[case("https://github.com/jwodder/headerparser/")]
+#[case("https://www.github.com/jwodder/headerparser")]
+#[case("http://github.com/jwodder/headerparser")]
+#[case("http://www.github.com/jwodder/headerparser")]
+#[case("github.com/jwodder/headerparser")]
+#[case("github.com/jwodder/headerparser.git")]
+#[case("github.com/jwodder/headerparser.git/")]
+#[case("github.com/jwodder/headerparser/")]
+#[case("www.github.com/jwodder/headerparser")]
+#[case("https://github.com/jwodder/none.git")]
+#[case("https://github.com/joe-coder/hello.world")]
+#[case("http://github.com/joe-coder/hello.world")]
+#[case("HTTPS://GITHUB.COM/joe-coder/hello.world")]
+#[case("HTTPS://WWW.GITHUB.COM/joe-coder/hello.world")]
+#[case("https://github.com/-Jerry-/geshi-1.0.git")]
+#[case("https://github.com/-Jerry-/geshi-1.0.git/")]
+#[case("https://github.com/-Jerry-/geshi-1.0/")]
+#[case("https://www.github.com/-Jerry-/geshi-1.0")]
+#[case("github.com/-Jerry-/geshi-1.0")]
+#[case("https://x-access-token:1234567890@github.com/octocat/Hello-World")]
+#[case("https://my.username@github.com/octocat/Hello-World")]
+#[case("https://user%20name@github.com/octocat/Hello-World")]
+#[case("https://user+name@github.com/octocat/Hello-World")]
+#[case("https://~user.name@github.com/octocat/Hello-World")]
+#[case("https://@github.com/octocat/Hello-World")]
+#[case("https://user:@github.com/octocat/Hello-World")]
+#[case("https://:pass@github.com/octocat/Hello-World")]
+#[case("https://:@github.com/octocat/Hello-World")]
+#[case("https://user:pass:extra@github.com/octocat/Hello-World")]
+fn test_try_from_bad_string(#[case] spec: String) {
+    match GHRepo::try_from(spec.clone()) {
         Err(ParseError::InvalidSpec(s)) if s == spec => (),
         e => panic!("Got wrong result: {e:?}"),
     }